Move Avatar and Rad into components
