+++ app/src/DesignSystem/index.js
@@ -1,9 +1,7 @@
-import Avatar from "./Avatar.svelte";
 import Button from "./Button.svelte";
 import Icon from "./Icon";
 import Input from "./Input";
 import Placeholder from "./Placeholder.svelte";
-import Rad from "./Rad.svelte";
 import Text from "./Text.svelte";
 
-export { Avatar, Button, Icon, Input, Placeholder, Rad, Text };
+export { Button, Icon, Input, Placeholder, Text };
