+++ app/src/pages/DesignSystem.svelte
@@ -1,7 +1,8 @@
 <script>
-  import { Button, Text, Icon, Input, Avatar } from "../DesignSystem";
+  import { Button, Text, Icon, Input } from "../DesignSystem";
 
   import Notification from "../components/Notification.svelte";
+  import Avatar from "../components/Avatar.svelte";
 
   import { link } from "svelte-spa-router";
