+++ app/src/pages/RegisterProject/TransactionSummaryStep.svelte
@@ -1,7 +1,9 @@
 <script>
   import { pop } from "svelte-spa-router";
   import TxRow from "../../components/TxRow.svelte";
-  import { Button, Text, Rad, Avatar } from "../../DesignSystem";
+  import { Button, Text } from "../../DesignSystem";
+  import Avatar from "../../components/Avatar.svelte";
+  import Rad from "../../components/Rad.svelte";
   import RowLayout from "../../layouts/helpers/Row.svelte";
 
   export let name = null;
